r trekker ut registrering-teller i egen funksjon
